main.py: Fix server_work() not running under gunicorn

+++ main.py
@@ -43,12 +43,18 @@ def landing():
 
 def server_work():
     "Run important server work before anything goes live"
+    if os.path.exists(".server-work.lock"):
+        logger.debug(".server-work.lock exists; skipping running server_work()")
+        return
+    # NOTE: 0644 is default
+    open(".server-work.lock", "w")
     logger.debug("Running git pull...")
     os.system("git pull")
     if os.system("which npx > /dev/null 2>&1") == 256:
         logger.debug("npx (npm) does not exist in the system path, so ./tscompile is "
                      "unrunnable. Aborting. Please install npm or make npx available "
                      "in the system path.")
+        os.remove(".server-work.lock")
         exit(1)
     if os.system("which tsc > /dev/null 2>&1") == 256:
         logger.debug("tsc does not exist in the system path; installing it globally...")
@@ -56,17 +62,22 @@ def server_work():
     logger.debug("Running ./tscompile...")
     if os.system("./tscompile") == 256:
         logger.debug("./tscompile failed to run, aborting.")
+        os.remove(".server-work.lock")
         exit(1)
+    logger.debug("All done.")
+    if os.path.exists(".server-work.lock"):
+        os.remove(".server-work.lock")
 
 
 def main():
     "Main function"
-    server_work()
     logger.debug("Serving landing.html")
     debug_mode = len(sys.argv) > 1 and sys.argv[1] == "+debug"
     app.run("0.0.0.0", debug=debug_mode)
     logger.debug("Program exited (0)")
 
 
+# server_work() must run when app is imported by another program
+server_work()
 if __name__ == "__main__":
     main()
